+++ src/main/java/ru/hogwarts/school/service/StudentService.java
@@ -47,11 +47,11 @@ public class StudentService {
         studentRepository.save(student);
         return studentDTO;
     }
-    public StudentDTO getStudentById(int id) {
+    public StudentDTO getStudentById(long id) {
         return fromStudent(studentRepository.findById(id));
     }
 
-    public void removeStudent(int id) {
+    public void removeStudent(long id) {
         studentRepository.deleteById(id);
     }
 
@@ -78,7 +78,7 @@ public class StudentService {
         return faculty;
     }
 
-    public List<StudentDTO> getStudentsByIdFaculty(int id) {
+    public List<StudentDTO> getStudentsByIdFaculty(long id) {
         List<Student> studentByFacultyId = studentRepository.findStudentByFacultyId(id);
         List<StudentDTO> studentDTOS = new ArrayList<>();
